Move email logic
