+++ mobile/screens/Onboarding/HandlerInformationScreen.tsx
@@ -14,6 +14,9 @@ import SolidDropDown from "../../components/SolidDropDown";
 import DateInput from "../../components/DateInput";
 import { validateBirthday, validateVisitDate } from "../../utils/helper";
 import { endOfExecutionHandler, ErrorWrapper } from "../../utils/error";
+import { sendEmail } from "../../../backend/server/utils/Authentication";
+import { EmailSubject } from "../../../backend/src/utils/types";
+import { EmailTemplate } from "../../../backend/src/utils/types";
 
 export default function HandlerInformationScreen(props: any) {
   const [dropDownValue, setDropDownValue] = useState("");
@@ -70,6 +73,19 @@ export default function HandlerInformationScreen(props: any) {
           default: "Failed to Update Handler Information",
         },
       });
+      console.log(user);
+      const emailData = {
+        email: (user as User).email,
+        firstName: firstName,
+        lastName: lastName,
+        address: address,
+      }
+      await sendEmail(
+        "gt.engineering@hack4impact.org",
+        EmailSubject.ACCOUNT_CREATED,
+        EmailTemplate.ACCOUNT_CREATED,
+        emailData
+      );
       return user;
     } catch (e) {
       endOfExecutionHandler(e as Error);
